requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
mysql-connector-python
python-dotenv
midtransclient
alembic
pytest
